requirements.txt
python-dotenv
pandas
numpy
scikit-learn
fastai